+++ cartridges/app_custom_cartridge/steptypes.json
@@ -2,15 +2,26 @@
     "step-types": {
         "script-module-step": [
             {
-                "@type-id": "custom.checkProductInfo",
+                "@type-id": "custom.productListExport",
                 "@support-parallel-execution": "false",
                 "@support-site-context": "true",
                 "@support-organization-context": "false",
-                "description": "Add products having information attr to a file",
-                "module": "app_custom_cartridge/cartridge/scripts/productList/checkPdtInfo.js",
+                "description": "export products having information attribute",
+                "module": "app_custom_cartridge/cartridge/scripts/productList/productListExport.js",
                 "function": "Process",
                 "transactional": "false",
-                "parameters": {},
+                "parameters": {
+                    "parameter":[
+                        {
+                            "@name": "exportFile",
+                            "@type": "string",
+                            "@required": "true",
+                            "@trim": "true",
+                            "description": "export the products contains the information attribute",
+                            "default-value": "exportlist_1"
+                        }     
+                    ]
+                },
                 "status-codes": {
                     "status": [
                         {
